apps/worker/hret_runner.py
"""HRET runner for executing evaluations."""

import json
import logging
import os
import subprocess
import tempfile
import time
from typing import Any, Dict, List, Optional
import yaml

logger = logging.getLogger(__name__)


class HRETRunner:
    """Runner for HRET evaluation toolkit."""
    
    def __init__(self, config_path: Optional[str] = None):
        """Initialize HRET runner."""
        self.config_path = config_path
        self.temp_dir = tempfile.mkdtemp(prefix="benchhub_plus_")
        logger.info(f"HRET runner initialized with temp dir: {self.temp_dir}")
    
    def run_evaluation(
        self,
        plan_yaml: str,
        models: List[Dict[str, Any]],
        timeout: int = 1800  # 30 minutes
    ) -> Dict[str, Any]:
        """Run evaluation using HRET."""
        
        try:
            # TODO: This is a placeholder implementation
            # In a real implementation, this would:
            # 1. Write plan.yaml to temp file
            # 2. Replace API keys in the plan
            # 3. Execute HRET with the plan
            # 4. Parse and return results
            
            logger.info("Starting HRET evaluation...")
            
            # Write plan to temporary file
            plan_file = os.path.join(self.temp_dir, "plan.yaml")
            with open(plan_file, "w", encoding="utf-8") as f:
                f.write(plan_yaml)
            
            # Replace API keys in plan
            self._inject_api_keys(plan_file, models)
            
            # Simulate HRET execution (placeholder)
            results = self._simulate_hret_execution(plan_file, models, timeout)
            
            logger.info("HRET evaluation completed successfully")
            return results
            
        except Exception as e:
            logger.error(f"HRET evaluation failed: {e}")
            raise
        finally:
            # Cleanup temporary files
            self._cleanup()
    
    def _inject_api_keys(self, plan_file: str, models: List[Dict[str, Any]]) -> None:
        """Inject API keys into plan file."""
        
        try:
            with open(plan_file, "r", encoding="utf-8") as f:
                plan_data = yaml.safe_load(f)
            
            # Create mapping of model names to API keys
            api_key_map = {model["name"]: model["api_key"] for model in models}
            
            # Update plan with actual API keys
            for model_config in plan_data.get("models", []):
                model_name = model_config.get("name")
                if model_name in api_key_map:
                    model_config["api_key"] = api_key_map[model_name]
            
            # Write updated plan back to file
            with open(plan_file, "w", encoding="utf-8") as f:
                yaml.dump(plan_data, f, default_flow_style=False, allow_unicode=True)
            
            logger.info("API keys injected into plan file")
            
        except Exception as e:
            logger.error(f"Failed to inject API keys: {e}")
            raise
    
    def _simulate_hret_execution(
        self,
        plan_file: str,
        models: List[Dict[str, Any]],
        timeout: int
    ) -> Dict[str, Any]:
        """Simulate HRET execution (placeholder implementation)."""
        
        # TODO: Replace this with actual HRET execution
        # This is a placeholder that simulates evaluation results
        
        logger.info("Simulating HRET execution...")
        
        # Simulate processing time
        time.sleep(2)
        
        # Load plan to get configuration
        with open(plan_file, "r", encoding="utf-8") as f:
            plan_data = yaml.safe_load(f)
        
        metadata = plan_data.get("metadata", {})
        sample_size = metadata.get("sample_size", 100)
        
        # Generate simulated results for each model
        results = {
            "metadata": {
                "plan_file": plan_file,
                "execution_time": 2.0,
                "sample_size": sample_size,
                "timestamp": time.time()
            },
            "model_results": []
        }
        
        for model in models:
            # Simulate evaluation results
            import random
            random.seed(42)  # For reproducible results
            
            # Simulate correctness scores
            correctness_scores = [random.uniform(0.3, 0.9) for _ in range(sample_size)]
            average_score = sum(correctness_scores) / len(correctness_scores)
            
            model_result = {
                "model_name": model["name"],
                "total_samples": sample_size,
                "correct_samples": sum(1 for score in correctness_scores if score > 0.5),
                "accuracy": sum(1 for score in correctness_scores if score > 0.5) / sample_size,
                "average_score": average_score,
                "execution_time": random.uniform(1.0, 3.0),
                "metadata": {
                    "api_base": model["api_base"],
                    "model_type": model.get("model_type", "openai")
                }
            }
            
            results["model_results"].append(model_result)
            
            # Generate sample-level results for database storage
            self._generate_sample_results(
                model["name"],
                correctness_scores,
                metadata
            )
        
        logger.info(f"Generated simulated results for {len(models)} models")
        return results
    
    def _generate_sample_results(
        self,
        model_name: str,
        correctness_scores: List[float],
        metadata: Dict[str, Any]
    ) -> List[Dict[str, Any]]:
        """Generate sample-level results for database storage."""
        
        sample_results = []
        
        for i, score in enumerate(correctness_scores):
            sample_result = {
                "prompt": f"Sample prompt {i+1} for {model_name}",
                "answer": f"Sample answer {i+1} from {model_name}",
                "skill_label": metadata.get("task_type", "QA"),
                "target_label": metadata.get("language", "English"),
                "subject_label": metadata.get("subject_type", "General"),
                "format_label": "text",
                "dataset_name": "benchhub_simulated",
                "meta_data": json.dumps({
                    "model_name": model_name,
                    "sample_index": i,
                    "simulated": True
                }),
                "correctness": score
            }
            sample_results.append(sample_result)
        
        # Store results in a temporary file for later processing
        results_file = os.path.join(self.temp_dir, f"{model_name}_samples.json")
        with open(results_file, "w", encoding="utf-8") as f:
            json.dump(sample_results, f, indent=2)
        
        return sample_results
    
    def _execute_hret_command(self, plan_file: str, timeout: int) -> subprocess.CompletedProcess:
        """Execute actual HRET command (placeholder)."""
        
        # TODO: Implement actual HRET command execution
        # This would be something like:
        # cmd = ["hret", "run", "--config", plan_file, "--output", output_file]
        # result = subprocess.run(cmd, timeout=timeout, capture_output=True, text=True)
        
        # For now, return a mock successful result
        return subprocess.CompletedProcess(
            args=["hret", "run", "--config", plan_file],
            returncode=0,
            stdout="Evaluation completed successfully",
            stderr=""
        )
    
    def _cleanup(self) -> None:
        """Clean up temporary files."""
        
        try:
            import shutil
            if os.path.exists(self.temp_dir):
                shutil.rmtree(self.temp_dir)
                logger.info(f"Cleaned up temp directory: {self.temp_dir}")
        except Exception as e:
            logger.warning(f"Failed to cleanup temp directory: {e}")
    
    def validate_plan(self, plan_yaml: str) -> bool:
        """Validate HRET plan configuration."""
        
        try:
            plan_data = yaml.safe_load(plan_yaml)
            
            # Basic validation
            required_keys = ["version", "metadata", "models", "datasets"]
            for key in required_keys:
                if key not in plan_data:
                    logger.error(f"Missing required key in plan: {key}")
                    return False
            
            # Validate models
            models = plan_data.get("models", [])
            if not models:
                logger.error("No models specified in plan")
                return False
            
            for model in models:
                if "name" not in model or "api_base" not in model:
                    logger.error("Model missing required fields (name, api_base)")
                    return False
            
            # Validate datasets
            datasets = plan_data.get("datasets", [])
            if not datasets:
                logger.error("No datasets specified in plan")
                return False
            
            logger.info("Plan validation successful")
            return True
            
        except yaml.YAMLError as e:
            logger.error(f"Invalid YAML in plan: {e}")
            return False
        except Exception as e:
            logger.error(f"Plan validation error: {e}")
            return False


def create_hret_runner() -> HRETRunner:
    """Factory function to create HRET runner."""
    return HRETRunner()